refactor(dotnet.NRI): add `propertyOverrides` to PackPackagesOptionsType
Allows MSBuild properties such as Version to be overridden via the `dotnet pack` with a fixed argument order.
Previously, the version was overridden to a dummy version by appending the necessary args to the output of GetPackCommand. This would result in MSBuild including those args as part of the value passed to the `-o` (output dir) parameter.