fix(dotnet.ZPD): fix placeholders PropertyGroup condition almost always evaluating to `true`
RepoRoot being the nearest directory containing "./.git/index".
Previously, the condition would evaluate to "true" because the MSBuildProjectDirectory would always be the full path of the directory  containing _ZipPublishDir.targets_ rather than the project that imported it. Because this is packaged in an NPM package named "@halospv3/hce.shared-config", the path would always contain "hce.shared".